python_requirements.txt
Django==1.8.5
MySQL-python==1.2.5
djangorestframework==3.1
djangorestframework-csv==1.3.4
django-filter==0.9.2
wsgiref==0.1.2
pycrypto==2.6.1
pytz>=2015b
python-dateutil==2.4.2
django-cors-headers==1.0.0
boto==2.36.0
fs==0.5.1
requests
uWSGI==2.0.11.2
Pygments==2.0.2
Markdown==2.6.3
